refactor: 拆分2048小游戏组件至不同模块
