+++ src/renderer/components/states/maxExpand/components/games/2048/config/constants.ts
@@ -1,0 +1,8 @@
+export const SIZE = 4;
+export const CELL = 64;
+export const GAP = 7;
+export const PAD = 8;
+export const BOARD = SIZE * CELL + (SIZE - 1) * GAP + 2 * PAD;
+export const SLIDE_MS = 120;
+
+export const STORAGE_KEY = 'island_game_2048_state';
