+++ src/renderer/components/states/maxExpand/components/games/2048/index.ts
@@ -1,0 +1,8 @@
+export { Game2048 } from './components/Game2048';
+export type {
+  Game2048EndPayload,
+  Game2048Handle,
+  Game2048Props,
+  Game2048Session,
+  Game2048State,
+} from './config/types';
